Renamed Graph Host Impl files
