+++ core/src/impl/Kokkos_Default_GraphNode_Impl.hpp
@@ -48,7 +48,7 @@
 #include <Kokkos_Macros.hpp>
 
 #include <impl/Kokkos_EBO.hpp>
-#include <impl/Kokkos_Host_Graph_fwd.hpp>
+#include <impl/Kokkos_Default_Graph_fwd.hpp>
 
 #include <Kokkos_Graph.hpp>
